requirements.txt
pytest~=8.3.3
pydantic_settings
requests
python-dotenv
selenium~=4.27.1
requests
selene>=2.0.0rc4
allure-pytest
appium-python-client~=4.4.0
jsonschema